Draft: Resolve "test new feature"

Closes #13

See merge request mygroup2593439/xfracod-slim!15

+++ include/S4.h
@@ -5,18 +5,18 @@
 class S4
 {
 public:
-	std::vector<float> b;
-	std::vector<float> d;
-	std::vector<float> b1;
-	std::vector<float> d0;
-	std::vector<float> df0;
-	std::vector<float> df;
-	std::vector<std::vector<float>> c_d;
-	std::vector<std::vector<float>> c_s;
-	std::vector<std::vector<float>> c;
+	std::vector<double> b;
+	std::vector<double> d;
+	std::vector<double> b1;
+	std::vector<double> d0;
+	std::vector<double> df0;
+	std::vector<double> df;
+	std::vector<std::vector<double>> c_d;
+	std::vector<std::vector<double>> c_s;
+	std::vector<std::vector<double>> c;
 
-	std::vector<float> b0;
-	std::vector<float> b0_old;
+	std::vector<double> b0;
+	std::vector<double> b0_old;
 	S4(int length); 
 	void limit_d();     // limit the d() value for stability
